Add Vista3D inference
Currently, since MONAI writes the result to the filesystem, the python
backend accepts the image from the frontend, writes the image to a
temporary file, and then runs the monai.bundle with the same python
environment that's being run by the server (so the poetry-installed deps
should work). The result is then read from a file and sent as a BLOB to
the frontend (as a `.nii.gz`). This should be decompressed and displayed
on the frontend.

- Add `2025-nvidiagtcdc/api.py` for storing all backend functions of
  models from NVIDIA collaboration
- Add store for Vista3D result as BLOB.
- Comment out server module and replace with Vista3D module,
  essentially a copy of the server module but with reduced and renamed
  functionality.
- Note, there are issues with the current approach once the BLOB gets to
  the server-side. A filesystem approach may be necessary. A filesystem
  approach would make the frontend and backend pinned to the same
  device.