Refactor MONAI segmentation to use VTK.js data pipeline
Refactor end-to-end MONAI segmentation process to use a VTK.js-native
data pipeline instead of file blobs. The Python backend now converts the
segmentation result directly into a vtk.js dictionary, which is sent to
the client. On the frontend, the data is deserialized into a
vtkImageData object and loaded as a new layer. This new layer is then
automatically converted into a labelmap and displayed on the base image.